+++ yt/yql/providers/yt/lib/access_provider/full/yt_access_provider.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <yt/yql/providers/yt/lib/access_provider/yt_access_provider.h>
+
+namespace NYql {
+
+IYtAccessProvider::TPtr CreateYtAccessProvider(const ITvmClient::TPtr& tvmClient, const TYtGatewayConfig& ytGatewayConfig);
+
+}; // namespace NYql