+++ yt/yql/providers/yt/provider/yql_yt_provider.cpp
@@ -448,9 +448,13 @@ TDataProviderInitializer GetYtNativeDataProviderInitializer(IYtGateway::TPtr gat
         info.Source = CreateYtDataSource(ytState);
         info.Sink = CreateYtDataSink(ytState);
         info.SupportFullResultDataSink = true;
-        info.OpenSession = [gateway, statWriter, qContext, fullCapture](const TString& sessionId, const TString& username,
+        info.OpenSession = [
+            gateway, statWriter, qContext, fullCapture, useSecureTmp = ytState->UseSecureTmp
+        ](
+            const TString& sessionId, const TString& username,
             const TOperationProgressWriter& progressWriter, const TYqlOperationOptions& operationOptions,
-            TIntrusivePtr<IRandomProvider> randomProvider, TIntrusivePtr<ITimeProvider> timeProvider) {
+            TIntrusivePtr<IRandomProvider> randomProvider, TIntrusivePtr<ITimeProvider> timeProvider
+        ) {
             gateway->OpenSession(
                 IYtGateway::TOpenSessionOptions(sessionId)
                     .UserName(username)
@@ -461,6 +465,7 @@ TDataProviderInitializer GetYtNativeDataProviderInitializer(IYtGateway::TPtr gat
                     .StatWriter(statWriter)
                     .QContext(qContext)
                     .FullCapture(fullCapture)
+                    .UseSecureTmp(useSecureTmp)
             );
             return NThreading::MakeFuture();
         };